+++ ui/local-malformed-1.rs
@@ -1,0 +1,16 @@
+#![no_main]
+
+#[mock::app]
+mod app {
+    #[shared]
+    struct Shared {}
+
+    #[local]
+    struct Local {}
+
+    #[task(local = [a:])]
+    fn foo(_: foo::Context) {}
+
+    #[init]
+    fn init(_: init::Context) -> (Shared, Local, init::Monotonics) {}
+}